backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1
httpx>=0.27.0
openai>=1.12.0
google-genai>=0.4.0
bcrypt>=4.1.2
pydantic-settings>=2.1.0
redis>=5.0.1
beautifulsoup4>=4.12.3
lxml>=5.1.0
python-multipart>=0.0.6
mcp[cli]>=1.3.0
cryptography>=42.0.2
sse-starlette>=2.0.0
numpy>=1.26.0
